ENH Updated default parameter settings and algorithms in Vessel Segmentation stuff
git-svn-id: http://svn.slicer.org/Slicer4@8938 3bd1e089-480b-0410-8dfb-8563597acbee

+++ branches/NCI/Slicer3.3-nci-test1/Applications/CLI/VesselnessMeasure.cxx
@@ -39,7 +39,7 @@ template<class T> int DoIt( int argc, char * argv[], T )
   PARSE_ARGS;
 
   //Pixel Types
-  typedef    short   InputPixelType;
+  typedef    double   InputPixelType;
   typedef    double   VesselnessPixelType;
 //  typedef    T       OutputPixelType;
 //  typedef    unsigned char       OutputPixelType;
